Update project schema; script populating projects
Projects now only contain a name, date (yyyy-mm), a flag to showcase the project on the home page, a list of images with URLs and captions, and an optional customer review.
Added scripting to dynamically populate projects on both the home and projects page. All projects are inserted on the projects page, whereas the index page will display a random project with the "isShowcased" attribute set as true.